agent/requirements.txt
browser-use
playwright
langchain-openai
supabase
cryptography
python-dotenv
